+++ cmd/zapi/cmd/new/new.go
@@ -6,9 +6,9 @@ import (
 	"fmt"
 
 	"github.com/brimsec/zq/api"
-	"github.com/brimsec/zq/archive"
 	"github.com/brimsec/zq/cmd/zapi/cmd"
 	"github.com/brimsec/zq/pkg/units"
+	"github.com/brimsec/zq/ppl/archive"
 	"github.com/mccanne/charm"
 )
