+++ ppl/archive/stat.go
