+++ ppl/archive/staticsource.go
