+++ ppl/archive/ztests/seek-index-overlap.yaml
@@ -9,7 +9,7 @@ script: |
 
 inputs:
   - name: babble.tzng
-    source: ../../ztests/suite/data/babble.tzng
+    source: ../../../ztests/suite/data/babble.tzng
 
 outputs:
   - name: stdout
