+++ ppl/archive/ztests/seek-index.yaml
@@ -7,7 +7,7 @@ script: |
 
 inputs:
   - name: babble.tzng
-    source: ../../ztests/suite/data/babble.tzng
+    source: ../../../ztests/suite/data/babble.tzng
 
 outputs:
   - name: stdout
