+++ ppl/cmd/zar/compact/command.go
@@ -5,7 +5,7 @@ import (
 	"flag"
 	"os"
 
-	"github.com/brimsec/zq/archive"
+	"github.com/brimsec/zq/ppl/archive"
 	"github.com/brimsec/zq/ppl/cmd/zar/root"
 	"github.com/mccanne/charm"
 )
