+++ ppl/cmd/zar/index/command.go
@@ -7,11 +7,11 @@ import (
 	"os"
 	"sync"
 
-	"github.com/brimsec/zq/archive"
 	"github.com/brimsec/zq/cli/procflags"
 	"github.com/brimsec/zq/field"
 	"github.com/brimsec/zq/pkg/rlimit"
 	"github.com/brimsec/zq/pkg/signalctx"
+	"github.com/brimsec/zq/ppl/archive"
 	"github.com/brimsec/zq/ppl/cmd/zar/root"
 	"github.com/mccanne/charm"
 )
